+++ sdk/python/pulumi_pagerduty/user_notification_rule.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 from . import outputs
 from ._inputs import *
@@ -18,7 +18,7 @@ class UserNotificationRule(pulumi.CustomResource):
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
                  contact_method: Optional[pulumi.Input[pulumi.InputType['UserNotificationRuleContactMethodArgs']]] = None,
-                 start_delay_in_minutes: Optional[pulumi.Input[float]] = None,
+                 start_delay_in_minutes: Optional[pulumi.Input[int]] = None,
                  urgency: Optional[pulumi.Input[str]] = None,
                  user_id: Optional[pulumi.Input[str]] = None,
                  __props__=None,
@@ -80,7 +80,7 @@ class UserNotificationRule(pulumi.CustomResource):
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[pulumi.InputType['UserNotificationRuleContactMethodArgs']] contact_method: A contact method block, configured as a block described below.
-        :param pulumi.Input[float] start_delay_in_minutes: The delay before firing the rule, in minutes.
+        :param pulumi.Input[int] start_delay_in_minutes: The delay before firing the rule, in minutes.
         :param pulumi.Input[str] urgency: Which incident urgency this rule is used for. Account must have the `urgencies` ability to have a low urgency notification rule. Can be `high` or `low`.
         :param pulumi.Input[str] user_id: The ID of the user.
         """
@@ -124,7 +124,7 @@ class UserNotificationRule(pulumi.CustomResource):
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
             contact_method: Optional[pulumi.Input[pulumi.InputType['UserNotificationRuleContactMethodArgs']]] = None,
-            start_delay_in_minutes: Optional[pulumi.Input[float]] = None,
+            start_delay_in_minutes: Optional[pulumi.Input[int]] = None,
             urgency: Optional[pulumi.Input[str]] = None,
             user_id: Optional[pulumi.Input[str]] = None) -> 'UserNotificationRule':
         """
@@ -135,7 +135,7 @@ class UserNotificationRule(pulumi.CustomResource):
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[pulumi.InputType['UserNotificationRuleContactMethodArgs']] contact_method: A contact method block, configured as a block described below.
-        :param pulumi.Input[float] start_delay_in_minutes: The delay before firing the rule, in minutes.
+        :param pulumi.Input[int] start_delay_in_minutes: The delay before firing the rule, in minutes.
         :param pulumi.Input[str] urgency: Which incident urgency this rule is used for. Account must have the `urgencies` ability to have a low urgency notification rule. Can be `high` or `low`.
         :param pulumi.Input[str] user_id: The ID of the user.
         """
@@ -159,7 +159,7 @@ class UserNotificationRule(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="startDelayInMinutes")
-    def start_delay_in_minutes(self) -> pulumi.Output[float]:
+    def start_delay_in_minutes(self) -> pulumi.Output[int]:
         """
         The delay before firing the rule, in minutes.
         """
